SAR/requirements.txt
certifi==2022.6.15
pip==21.2.4
setuptools==61.2.0
six==1.16.0
typing_extensions==4.1.1
wheel==0.37.1
absl-py==1.1.0
accelerate==0.10.0
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.2
atpublic==3.0.1
attrs==21.4.0
babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
bert-score==0.3.11
bleach==5.0.0
cachetools==5.2.0
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
cycler==0.11.0
datasets==2.3.2
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.5.1
entrypoints==0.4
executing==0.8.3
faiss-cpu==1.7.2
fastjsonschema==2.15.3
filelock==3.7.1
flufl-lock==7.0
fonttools==4.33.3
frozenlist==1.3.0
fsspec==2022.5.0
google-auth==2.8.0
google-auth-oauthlib==0.4.6
grpcio==1.46.3
hjson==3.1.0
huggingface-hub==0.7.0
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.8.0
ipykernel==6.15.0
ipython==8.4.0
ipython-genutils==0.2.0
jedi==0.18.1
jinja2==3.1.2
joblib==1.1.0
json5==0.9.8
jsonschema==4.6.0
jupyter-client==7.3.4
jupyter-core==4.10.0
jupyter-server==1.17.1
jupyterlab==3.4.3
jupyterlab-pygments==0.2.2
jupyterlab-server==2.14.0
kiwisolver==1.4.3
lxml==4.9.1
markdown==3.3.7
markupsafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mauve-text==0.3.0
mistune==0.8.4
multidict==6.0.2
multiprocess==0.70.13
nbclassic==0.3.7
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
ninja==1.10.2.3
nltk==3.7
notebook==6.4.12
notebook-shim==0.1.0
numpy==1.22.4
oauthlib==3.2.0
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==9.1.1
portalocker==2.5.1
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==8.0.0
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.2
pygments==2.12.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
pyyaml==6.0
pyzmq==23.1.0
regex==2022.6.2
requests==2.28.0
requests-oauthlib==1.3.1
responses==0.18.0
rouge==1.0.0
rsa==4.8
sacrebleu==2.2.1
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
send2trash==1.8.0
sentencepiece==0.1.96
sniffio==1.2.0
soupsieve==2.3.2.post1
stack-data==0.3.0
tabulate==0.9.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardx==2.5.1
termcolor==1.1.0
terminado==0.15.0
threadpoolctl==3.1.0
tinycss2==1.1.1
tokenizers==0.12.1
tornado==6.1
tqdm==4.64.0
traitlets==5.3.0
transformers==4.20.0
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
werkzeug==2.1.2
xxhash==3.0.0
yacs==0.1.8
yarl==1.7.2
zipp==3.8.0
wandb